+++ test/stdvector_serialize.cpp
@@ -14,7 +14,7 @@ int stdvector_serialize(int, char *[])
     VERIFY(json.isArray());
 
     QJsonArray array(json.toArray());
-    VERIFY(array.size() == v.size());
+    VERIFY(static_cast<size_t>(array.size()) == v.size());
     VERIFY(array.at(0).toInt() == v[0]);
     VERIFY(array.at(1).toInt() == v[1]);
     VERIFY(array.at(2).toInt() == v[2]);
